+++ main.py
@@ -1,5 +1,5 @@
 
-from src.App import main
+from src.fApp import main
 from src.ConsoleLogger import logger
 
 r"""
